+++ packages/embedded-client/src/index.ts
@@ -1,9 +1,3 @@
-/**
- * Client library for interacting with embedded Cerbos policy decision points from server-side Node.js and browser-based applications.
- *
- * @packageDocumentation
- */
-
 export * from "./client.js";
 export * from "./loader.js";
 export * from "./options.js";