+++ packages/hub/src/types/File.ts
@@ -1,7 +1,5 @@
 /**
  * A file in a store.
- *
- * @public
  */
 export interface File {
   /**
@@ -14,7 +12,7 @@ export interface File {
    *
    * @remarks
    * Use {@link bytesToString} and {@link stringToBytes} to convert the contents to or from a string, and
-   * {@link @cerbos/files#parsePolicy} and {@link @cerbos/files#serializePolicy} to programmatically manipulate policies.
+   * {@link @cerbos/files!parsePolicy | parsePolicy} and {@link @cerbos/files!serializePolicy | serializePolicy} to programmatically manipulate policies.
    */
   contents: Uint8Array;
 }