+++ packages/hub/src/types/GetFilesRequest.ts
@@ -1,7 +1,5 @@
 /**
  * Input to {@link StoresClient.getFiles}.
- *
- * @public
  */
 export interface GetFilesRequest {
   /**