+++ packages/react/src/index.ts
@@ -1,13 +1,2 @@
-/**
- * A collection of React hooks for interacting with Cerbos policy decision points.
- *
- * @packageDocumentation
- */
-export { type CerbosProviderProps, CerbosProvider } from "./cerbos-provider.js";
-export { useCerbos } from "./use-cerbos.js";
-export {
-  type AsyncResult,
-  useCheckResource,
-  useCheckResources,
-  useIsAllowed,
-} from "./use-cerbos-request.js";
+export * from "./provider.js";
+export * from "./hooks.js";